Bug Fixes: * Follow up 2d2991a, remove trailing spaces in inter_athena.conf * Added `bound` field on `picklog` table (bugreport:9240) Thank @icxbb-xx (Napster) * Fixed Randomize Spell infinite loop. Failure of skill improvise db reading. (bugreport:9288) * Added **max_extended_aspd** for caping max ASPD for KO/Rebellion, default is 193. Pointed by @raynra & Ziu * Mail attachment   * Added `bound` field on `mail` table, since group with PC_PERM_TRADE_BOUNDED can put bound item into attachment   * Fixed wrong data type causing wrong zeny amount (bugreport:9291) * Fixed item color issue for bound item (bugreport:9238)   * Stackable item shouldn't be displayed as yellow color   * Equip doesn't shown as yellow after relog or item is moved * Silence status immunity for Renewal (bugreport:9227)   * Bleeding (SC_BLEEDING): Vit -> Agi   * Silence (SC_SILENCE): Vit -> Int   * Sleep (SC_SLEEP): Int -> Agi   * Thank @Playtester, @Lilith-, partial merge of Hercules:6f26451 * Fixed status check when cast a skill to hidden target   * Stone Curse shouldn't be casted to hidden enemy when the caster has intravision (bugreport:9266)   * Added **INF3_HIT_HIDING** for skills that *maybe* have effect to/can hit the hidden target. Previously, every Earth Element skills *always* had it, doesn't matter ground skill or single-target skill   * **!TODO: Confirm other skills that have ability for this!**   * **!FIXME: For now, I rely as the previous check did, I put some Earth Element skills (and) with area/ground effect** * Correcting Devotion behavior (Hercules:e7be725, 41f28c0)   * Failing to cast Devotion to the 6th player   * Only player who devotes other that 'shout' Auto Guard when it active also with the walk delay   * Status icon that inherited from devoter won't be displayed at devoted player   * Reflect Damage on devoted player won't reflect normal attack, only melee skill attack   * Added config **devotion_rdamage_skill_only** for backward compability, ignore the behavior above
**IMPORTANT**
  * Please import sql-files/upgrades/upgrade_20140915_log.sql
  * Please import sql-files/upgrades/upgrade_20140915.sql

+++ src/map/mail.c
@@ -64,7 +64,7 @@ int mail_removezeny(struct map_session_data *sd, short flag)
 * @param amount
 * @return True if item/zeny can be set, False if failed
 */
-bool mail_setitem(struct map_session_data *sd, short idx, unsigned short amount) {
+bool mail_setitem(struct map_session_data *sd, short idx, int amount) {
 
 	if( pc_istrading(sd) )
 		return false;
